FIX #47 : 바코드 인식 안되는 문제 해결
