+++ Community/CommunityTab.swift
@@ -13,7 +13,6 @@ import SnapKit
 class CommunityTab: UIViewController {
     
     let mainView = CommunityTabView()
-    let createCommunityVC = CreateCommunityViewController()
     
     var communities = ((UIApplication.shared.delegate as? AppDelegate)?.communities)!
 
@@ -48,10 +47,11 @@ class CommunityTab: UIViewController {
     }
     
     @objc func searchButtonPress() {
-        self.navigationController?.pushViewControllerTabHidden(SearchCommunity(), animated: true)
+        self.navigationController?.pushViewControllerTabHidden(SearchCommunityViewController(), animated: true)
     }
 
     @objc func addButtonPress() {
+        let createCommunityVC = CreateCommunityViewController()
         self.navigationController?.pushViewControllerTabHidden(createCommunityVC, animated: true)
     }
